+++ projects/dts-backoffice-util/src/lib/dts-backoffice-util.module.ts
@@ -1,11 +1,8 @@
 import { NgModule, ModuleWithProviders } from '@angular/core';
-import {
-  TotvsScheduleExecutionComponent
-} from './components/totvs-schedule-execution/totvs-schedule-execution.component';
+import { TotvsScheduleExecutionComponent } from './components/totvs-schedule-execution/totvs-schedule-execution.component';
 import { FormsModule } from '@angular/forms';
-import { PoModule, PoI18nService } from '@po-ui/ng-components';
-import { HttpClientModule } from '@angular/common/http';
-// import { BrowserModule } from '@angular/platform-browser';
+import { PoModule } from '@po-ui/ng-components';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { TotvsScheduleExecutionService } from './components/totvs-schedule-execution/totvs-schedule-execution.service';
 import { CommonModule } from '@angular/common';
 import { DtsDateFormatPipe } from './pipes/dts-date-format.pipe';
@@ -15,18 +12,18 @@ import { DtsDateFormatPipe } from './pipes/dts-date-format.pipe';
     TotvsScheduleExecutionComponent,
     DtsDateFormatPipe
   ],
-  imports: [
-    CommonModule,
-    PoModule,
-    FormsModule,
-    HttpClientModule
-  ],
   exports: [
     TotvsScheduleExecutionComponent,
     DtsDateFormatPipe
   ],
+  imports: [
+    CommonModule,
+    PoModule,
+    FormsModule
+  ],
   providers: [
-    TotvsScheduleExecutionService
+    TotvsScheduleExecutionService,
+    provideHttpClient(withInterceptorsFromDi())
   ]
 })